+++ openstack/blockstorage/v3/snapshots/delete.go
@@ -1,0 +1,8 @@
+package snapshots
+
+import "github.com/opentelekomcloud/gophertelekomcloud"
+
+func Delete(client *golangsdk.ServiceClient, id string) (err error) {
+	_, err = client.Delete(client.ServiceURL("snapshots", id), nil)
+	return
+}
