--- openstack/blockstorage/v3/snapshots/requests.go
@@ -1,189 +1,0 @@
-package snapshots
-
-import (
-	"github.com/opentelekomcloud/gophertelekomcloud"
-	"github.com/opentelekomcloud/gophertelekomcloud/pagination"
-)
-
-// CreateOptsBuilder allows extensions to add additional parameters to the
-// Create request.
-type CreateOptsBuilder interface {
-	ToSnapshotCreateMap() (map[string]interface{}, error)
-}
-
-// CreateOpts contains options for creating a Snapshot. This object is passed to
-// the snapshots.Create function. For more information about these parameters,
-// see the Snapshot object.
-type CreateOpts struct {
-	VolumeID    string            `json:"volume_id" required:"true"`
-	Force       bool              `json:"force,omitempty"`
-	Name        string            `json:"name,omitempty"`
-	Description string            `json:"description,omitempty"`
-	Metadata    map[string]string `json:"metadata,omitempty"`
-}
-
-// ToSnapshotCreateMap assembles a request body based on the contents of a
-// CreateOpts.
-func (opts CreateOpts) ToSnapshotCreateMap() (map[string]interface{}, error) {
-	return golangsdk.BuildRequestBody(opts, "snapshot")
-}
-
-// Create will create a new Snapshot based on the values in CreateOpts. To
-// extract the Snapshot object from the response, call the Extract method on the
-// CreateResult.
-func Create(client *golangsdk.ServiceClient, opts CreateOptsBuilder) (r CreateResult) {
-	b, err := opts.ToSnapshotCreateMap()
-	if err != nil {
-		r.Err = err
-		return
-	}
-	_, r.Err = client.Post(createURL(client), b, &r.Body, &golangsdk.RequestOpts{
-		OkCodes: []int{202},
-	})
-	return
-}
-
-// Delete will delete the existing Snapshot with the provided ID.
-func Delete(client *golangsdk.ServiceClient, id string) (r DeleteResult) {
-	_, r.Err = client.Delete(deleteURL(client, id), nil)
-	return
-}
-
-// Get retrieves the Snapshot with the provided ID. To extract the Snapshot
-// object from the response, call the Extract method on the GetResult.
-func Get(client *golangsdk.ServiceClient, id string) (r GetResult) {
-	_, r.Err = client.Get(getURL(client, id), &r.Body, nil)
-	return
-}
-
-// ListOptsBuilder allows extensions to add additional parameters to the List
-// request.
-type ListOptsBuilder interface {
-	ToSnapshotListQuery() (string, error)
-}
-
-type ListOpts struct {
-	// AllTenants will retrieve snapshots of all tenants/projects.
-	AllTenants bool `q:"all_tenants"`
-
-	// Name will filter by the specified snapshot name.
-	Name string `q:"name"`
-
-	// Status will filter by the specified status.
-	Status string `q:"status"`
-
-	// TenantID will filter by a specific tenant/project ID.
-	// Setting AllTenants is required to use this.
-	TenantID string `q:"project_id"`
-
-	// VolumeID will filter by a specified volume ID.
-	VolumeID string `q:"volume_id"`
-
-	// Comma-separated list of sort keys and optional sort directions in the
-	// form of <key>[:<direction>].
-	Sort string `q:"sort"`
-
-	// Requests a page size of items.
-	Limit int `q:"limit"`
-
-	// Used in conjunction with limit to return a slice of items.
-	Offset int `q:"offset"`
-
-	// The ID of the last-seen item.
-	Marker string `q:"marker"`
-}
-
-// ToSnapshotListQuery formats a ListOpts into a query string.
-func (opts ListOpts) ToSnapshotListQuery() (string, error) {
-	q, err := golangsdk.BuildQueryString(opts)
-	if err != nil {
-		return "", err
-	}
-	return q.String(), err
-}
-
-// List returns Snapshots optionally limited by the conditions provided in
-// ListOpts.
-func List(client *golangsdk.ServiceClient, opts ListOptsBuilder) pagination.Pager {
-	url := listURL(client)
-	if opts != nil {
-		query, err := opts.ToSnapshotListQuery()
-		if err != nil {
-			return pagination.Pager{Err: err}
-		}
-		url += query
-	}
-	return pagination.NewPager(client, url, func(r pagination.PageResult) pagination.Page {
-		return SnapshotPage{pagination.LinkedPageBase{PageResult: r}}
-	})
-}
-
-// UpdateMetadataOptsBuilder allows extensions to add additional parameters to
-// the Update request.
-type UpdateMetadataOptsBuilder interface {
-	ToSnapshotUpdateMetadataMap() (map[string]interface{}, error)
-}
-
-// UpdateMetadataOpts contain options for updating an existing Snapshot. This
-// object is passed to the snapshots.Update function. For more information
-// about the parameters, see the Snapshot object.
-type UpdateMetadataOpts struct {
-	Metadata map[string]interface{} `json:"metadata,omitempty"`
-}
-
-// ToSnapshotUpdateMetadataMap assembles a request body based on the contents of
-// an UpdateMetadataOpts.
-func (opts UpdateMetadataOpts) ToSnapshotUpdateMetadataMap() (map[string]interface{}, error) {
-	return golangsdk.BuildRequestBody(opts, "")
-}
-
-// UpdateMetadata will update the Snapshot with provided information. To
-// extract the updated Snapshot from the response, call the ExtractMetadata
-// method on the UpdateMetadataResult.
-func UpdateMetadata(client *golangsdk.ServiceClient, id string, opts UpdateMetadataOptsBuilder) (r UpdateMetadataResult) {
-	b, err := opts.ToSnapshotUpdateMetadataMap()
-	if err != nil {
-		r.Err = err
-		return
-	}
-	_, r.Err = client.Put(updateMetadataURL(client, id), b, &r.Body, &golangsdk.RequestOpts{
-		OkCodes: []int{200},
-	})
-	return
-}
-
-// IDFromName is a convienience function that returns a snapshot's ID given its name.
-func IDFromName(client *golangsdk.ServiceClient, name string) (string, error) {
-	count := 0
-	id := ""
-
-	listOpts := ListOpts{
-		Name: name,
-	}
-
-	pages, err := List(client, listOpts).AllPages()
-	if err != nil {
-		return "", err
-	}
-
-	all, err := ExtractSnapshots(pages)
-	if err != nil {
-		return "", err
-	}
-
-	for _, s := range all {
-		if s.Name == name {
-			count++
-			id = s.ID
-		}
-	}
-
-	switch count {
-	case 0:
-		return "", golangsdk.ErrResourceNotFound{Name: name, ResourceType: "snapshot"}
-	case 1:
-		return id, nil
-	default:
-		return "", golangsdk.ErrMultipleResourcesFound{Name: name, Count: count, ResourceType: "snapshot"}
-	}
-}
