--- openstack/blockstorage/v3/snapshots/testing/doc.go
@@ -1,2 +1,0 @@
-// snapshots_v3
-package testing
