--- openstack/blockstorage/v3/volumes/testing/doc.go
@@ -1,2 +1,0 @@
-// volumes_v3
-package testing
