--- openstack/blockstorage/v3/volumetypes/testing/doc.go
@@ -1,2 +1,0 @@
-// volume_types
-package testing
